+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Add first version of a cover/bare soil marker (#168, #200)
 - Avoid high committed memory for zonal stats calculation (#197)
 - General small improvements, e.g. save randomforest models compressed,.. (#144)
+- Use Exactextract as default engine for zonalstats calculation (#181)
 
 ### Bugs fixed